Amended the Scaff for what's actually available
1.5m lengths of scaff are no longer available, which is - to use the
technical vernacular - a huge pain in the arse. To compensate for this,
the entrance to the arena has been rejigged to use 1.25m sections (which
are available) and replacing one of the 2.5m sections with a 3m section.

This is a 25cm decrease in the size of the entrance, which is mildly
irritating, but has the advantage over the previous configuration of
actually being possible.

Signed-off-by: Alistair Lynn <alynn@studentrobotics.org>
Signed-off-by: Chris Kirkham <chrisjameskirkham@gmail.com>