Massage some scales.
It seems that tables are different sizes on different designs here, to
the extent that copy and pasting the socket object between designs
results in a different size relative to the table size.

So, reduce the size of the socket object on l2 and l3 so that they're
all roughly the same size in proportion to the table size.