Change the DSQ handling to a checkbox, add a 'present' field
While it would be nice to simply not record the TLA if a team isn't
present, it seems quite likely that this won't happen (since
linesmen tend to, incorrectly, get the TLA from the screens not
the robots). As a result, we need an explicit marker for the appearance
of a team in the arena.

Change-Id: I805bbbd7d34bcb93dcd2b5b1b64661349f458c19