Implement direction-to-top validation
This appears to mostly work, based on the (correct) token I have to
hand. There appear to be some cases where three markers are in sight
and a (false) error is found. In such cases, the angle between one
pair of markers is in the region 30-35 degrees, which is quite a
large error.

Change-Id: I5721edf8bb804a2f7bbd0f023c047d77173ab8de