Up tolerance to 35 degrees
Markers which are out of place should return values which are 90
degrees out, so in theory our tolerance could be up to 45. I'm not
keen on a value that large since the chance of missing failures is
quite large. Experimentation (using the sr-token-cube visualisation)
shows that invalid markers are typically in the 60+ degrees range,
so I think allowing 35 degrees between things which are 'parallel'
is reasonable.

Change-Id: I1fd2ad961b266f7bccd353d3270550753612debd