Improve marker direction estimates by using both edges
This has the functions which get a direction from a single marker
use both edges of the marker to generate the direction vector. This
should improve the estimates and reduce the comparative error.

Testing (using the sr-token-cube) shows that this improves things
in some cases, though not all, and doesn't appear to have any
negative impact.

Change-Id: I5f0ae7b1ebe6102ec16292dc7f68010d7b69f32f