Improve direction checking output
Change how the direction comparisons are done so that we can store
the results and use the new structures to greatly improve the errors
produced on the fly. Eventually I'd like to use this to be able to
tell the user which marker is at fault, rather than merely that one
of them is.

Change-Id: I6672e858267e22423c2020b11ec892da5a7ec066